third_party/blink/web_tests/fast/events/event-on-culled_inline-expected.txt

Click the empty area on the left
Click the empty areain middle

Click event should be fired when it occurs within culled inline.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.elementFromPoint(x, y).id is "clickme1"
LABEL that is a parent of INPUT is clicked.
LABEL that is a parent of INPUT is clicked.
PASS gotClick is true


PASS document.elementFromPoint(x, y).id is "clickme2"
SPAN that is a parent of EM is clicked.
PASS gotClick is true


PASS document.getElementById("checkbox").checked is true
PASS successfullyParsed is true

TEST COMPLETE
